+++ Documentation/changelog.md
@@ -2,6 +2,7 @@
 * New calculation system for thermistors
 * Add command M305 - Set thermistor and ADC parameters and DHT sensor parameters
 * Unified commands M320 and M420 in M420 for all Bed Level
+* Add subcode to M106 F<Fanpin> L<min speed> I<inverted logic>
 * Add EMERGENCY_PARSER to Due
 * Fix and clear code